Moved Delete handler to parent component and also added delete button on task modal
