+++ src/components/TaskManager/TaskManager.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import Home from "../Home/Home";
-import { Routes, Route, Navigate } from "react-router-dom";
+import { Routes, Route, Navigate, useNavigate } from "react-router-dom";
 import Login from "../Login/Login";
 import Register from "../Register/Register";
 import Dashboard from "../Dashboard/Dashboard";
@@ -10,11 +10,12 @@ import api from "../../api/api";
 import EditTask from "../EditTask/EditTask";
 import NewTask from "../NewTask/NewTask";
 import Stats from "../Stats/Stats.jsx";
-import { getTasksResponse, renderToast, updateTaskStatusApi } from "../../utils/utils";
+import { getTasksResponse, renderToast, updateTaskStatusApi, handleTaskDeleteResponse } from "../../utils/utils";
 import Spinner from "../Spinner/Spinner";
 import useAuthContext from "../../hooks/useAuthContext";
 
 const TaskManager = () => {
+  const navigate = useNavigate();
   const [tasks, setTasks] = useState([]);
   const { setAuth, authLoader, setAuthLoader, setUser, loading, setLoading } = useAuthContext();
 
@@ -61,6 +62,22 @@ const TaskManager = () => {
     }
   };
 
+  const handleTaskDelete = async (id) => {
+    setLoading(true);
+    try {
+      const response = await handleTaskDeleteResponse(id);
+      if (response.status === 200) {
+        renderToast("Task Deleted", "error");
+        navigate("/dashboard/tasks");
+        setTasks(tasks.filter((task) => task.id !== id));
+        setLoading(false);
+      }
+    } catch (err) {
+      console.log(err);
+      setLoading(false);
+    }
+  };
+
   if (authLoader) return <Spinner />;
 
   return (
@@ -72,8 +89,8 @@ const TaskManager = () => {
         <Route element={<PrivateRoutes />}>
           <Route path="/dashboard" element={<Dashboard />}>
             <Route path="/dashboard/stats" element={<Stats tasks={tasks} updateTaskStatus={updateTaskStatus} />} />
-            <Route path="/dashboard/tasks" element={<Tasks getTasks={getTasks} loading={loading} tasks={tasks} updateTaskStatus={updateTaskStatus} />} />
-            <Route path="/dashboard/tasks/edit/:id" element={<EditTask />} />
+            <Route path="/dashboard/tasks" element={<Tasks getTasks={getTasks} loading={loading} tasks={tasks} updateTaskStatus={updateTaskStatus} handleTaskDelete={handleTaskDelete} />} />
+            <Route path="/dashboard/tasks/edit/:id" element={<EditTask handleTaskDelete={handleTaskDelete} />} />
             <Route path="/dashboard/tasks/new" element={<NewTask />} />
           </Route>
         </Route>
